server/requirements.txt
Flask==2.2.3
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-SQLAlchemy==3.0.3
SQLAlchemy==2.0.3
psycopg2-binary==2.9.5
mysqlclient==2.1.1
bcrypt==4.0.1
Pillow==9.4.0
email-validator==1.3.1
python-dotenv==1.0.0
Werkzeug==2.2.3
alembic==1.10.4 